xandikos: change uid/gid within container to 0
the reason behind this change is that libvirtiofs for the vmdisk already
has a layer of uid/gid mapping which allocates a few tens of thousands
of uid/gids to each VM. this means that at least in vmdisk, it's
impossible to further use uid/gid mapping for podman.

this change makes xandikos run as root within its container.
effectively, this makes the container run as the xandikos user within
the service VM; if there is a container escape, it will be limited to
that. should be fine right?