matrix: update subdomain
we're using a subdomain because we use cname with dyndns,
which can't be used on the apex domain.
meanwhile, this change is because i believe other servers think that
this new server is impersonating our old synapse, and i have no idea how
to modify the signing key on this server to match the old one. so just
ditch the old name